Build INAppStoreWindow static library
This library does raise a non-fatal exception on Yosemite due to Apple
changing the way NSWindow works, but it does still work at the moment.

If a future update to OS X breaks INAppStoreWindow we'll find another
way to accomplish some of the things it provides that we're using, like
control over the way NSWindow top bars are rendered.