Build an Info.plist into the askpass binary with same ID as the main app.
This follows Apple's official guideline for embedded tools like this.

Despite the general rule that bundle IDs should be unique for every
piece of code, the bundle ID has to be the same in our askpass to allow
keychain operations to work by default.

Without using the same ID, the user will get a confusing keychain access
prompt no matter how the code is signed.