Don't report account status errors to the telemetry API
These are almost always going to be network issues or transient issues.

In the rare case that the network and API server are fine, it's going to
end up being something like a session token expiring which isn't a fatal
error either but will be better handled elsewhere in the codebase.