[App] only show open file window on repeat failure
Before, we only tried to unmount once, and if that failed we
immediately opened the open file warning window, which may be
inaccurate because the drive hasn’t truly had a chance to unmount. In
that case, applications may be attempting to close the open files and
we should give them a chance to finish.

Instead we will retry 5 times, with exponential backoff each time. If
the drive fails to unmount after 5 attempts, we then show the open
file window.

This may still be inaccurate in situations where there are a lot of
pending writes to an open file, but there’s nothing further we can do
about it.