[App] fix UI deadlock caused by network outage
If the network stops responding, the mountDetails property will still
continue asking the OS for filesystem details, and the OS will ask
our filesystem, which can’t respond because the network is unavailable.

Because these calls were happening in the main thread, they also
blocked the thread, which blocked UI rendering and events.

The immediate fix is to run the check on the background thread, however
we will still need to make sure the mount check thread itself doesn’t
block for very long, or at least unblocks once the network issue is
resolved