[App] send MountEvent object in volumeShouldMount
Used to tell the MountController whether a specific mount request is
user-initiated or automated, like a retry or a keepMounted triggered
mount.

The difference is that the MountController will lower the retry
interval to 5 seconds instead of 60, and will start a retry counter
to rapidly retry the mount if it fails temporarily. This ensures that
recoverable failures don’t tigger a flood of desktop notifications.

For non user-initiated mounts, we don’t want to tigger desktop
notifications at all, as the user isn’t waiting for or expecting
immediate feedback, and we don’t want to rapidly retry mounting if
the user isn’t waiting for the mount to finish anyway. In the worst
case, retrying every 5 seconds automatically might hammer the server
and even cause a “thundering herd” problem if a large number of clients
all line up and try at the same time.