[App] fix Swift compiler change for division
This is likely a result of a change in the way the Swift compiler
infers the result of division, both current and total are integers
but the result was always zero both before and after the cast to a
Double. The intermediate result has to be calculated with two
Doubles instead.