[App] add current SSH host keys to known_hosts
These are only used for rsync, as the SFTPFS code handles host key
checking internally and has never used the ssh binary at all.

We’re going to start checking them when connecting with rsync for an
unencrypted sync or restore. We should probably use the fingerprint
API, however that returns an md5 which can’t be reversed back to the
format that known_hosts requires, so we would have to add the base64
version to the fingerprint API and potentially calculate the md5 from
that, since SFTPFS (really SBB) requires that format instead of
base64.