[App] fix constraint layout quirks in sync panel
The older constraints being applied to elements in this panel may have
worked fine when compiled against older versions of the macOS SDK
and/or run on older versions of macOS, but as of the macOS 10.13 or
macOS 10.14 update something changed and caused the constraints to
stop producing the same layout, leading to elements drifting up out
of their intended location.

The constraint engine may have become more strict, but it’s a good
idea to be more specific either way, so we now add vertical centering
constraints to the buttons too.