[App] disable rsync xattr support temporarily
The system binary for rsync supports xattrs, but is extremely old and
causes other problems, which is why we use a bundled version of rsync
instead. However the bundled version requires a patch to support
xattrs on macOS, and the patch currently seems to cause a segfault in
rsync, which causes unencrypted sync to fail.

Until we can investigate the rsync segfault we will disable xattrs,
they aren’t a significant benefit anyway for most users, macOS uses
them for things like quarantine flags, downloads from browsers (to
embed the source URL to show users where it came from when quarantine
is being lifted), and various other things. Users don’t rely on them
directly.